refactor : lottogame interface 생성
