+++ src/main/java/step2/LottoApplication.java
@@ -12,9 +12,12 @@ public class LottoApplication {
         int ticketAmount = InputView.buyTicket();
         int manualLottoCount = InputView.buyManual();
         Ticket ticket = new Ticket(ticketAmount, manualLottoCount);
+
         List<String> manualNumbers = InputView.buyManual(manualLottoCount);
         ticket.buyManuals(manualNumbers);
+
         ResultView.printTickets(ticket);
+
         WinningLotto winningLottoNumber
                 = new WinningLotto(InputView.inputWinningNumbers(), InputView.inputBonusNumber());
         ResultView.printAnalysis(ticket, winningLottoNumber);
